+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Bug fixes:
+  - Remove module dependency on `PSRule`. [#13](https://github.com/microsoft/PSRule.Monitor/issues/13)
+
 ## v0.1.0-B1912005 (pre-release)
 
 - Initial pre-release.